backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==3.17.1
openai==1.3.0
chromadb==0.4.15
python-dotenv==1.0.0 